feat: don't use memory caches for ES
Memory caches depend on the projections for the UI api. There's a real
risk that the ES projections will use the memory caches at a point where
the memory cache is not in a good state (rebuilds, bugs, ...).

Instead we introduce a separate cache used by ES projections runner,
to be run before running any other ES projections so we know it's
up to date.